+++ blockfile/blk_file.h
@@ -14,13 +14,13 @@ class BlockFile
 public:
    FILE* fp;			// os file pointer
    char* filename;
-   int blocklength;	    // length of a block
-   int act_block; 	    // block # of fp's position (fp can stay at block boundaries)
+   size_t blocklength;	    // length of a block
+   size_t act_block; 	    // block # of fp's position (fp can stay at block boundaries)
    size_t number;		    // total # of blocks
    bool new_flag;		// specifies if this is a new file
 
      //=====================================================
-   BlockFile(char* name, int b_length);
+   BlockFile(char* name, size_t b_length);
    			        // Filename und Blocklaenge (fuer neue Files)
    ~BlockFile();
 
@@ -38,29 +38,29 @@ public:
 	      for (int i = 0; i < num; i ++)
 		  { *add = fgetc(fp); add ++; }*/}
 
-   void fwrite_number(int num);
+   void fwrite_number(size_t num);
 
-   int fread_number();
+   size_t fread_number();
 
-   void seek_block(int bnum)
+   void seek_block(size_t bnum)
 		{ fseek(fp,(bnum-act_block)*blocklength,SEEK_CUR); }
 
    void read_header(char * header);
 
    void set_header(char* header);
 
-   bool read_block(Block b,int i);
+   bool read_block(Block b,size_t i);
 
-   bool write_block(Block b,int i);
+   bool write_block(Block b,size_t i);
 
-   int append_block(Block b);
+   size_t append_block(Block b);
 
-   bool delete_last_blocks(int num);
+   bool delete_last_blocks(size_t num);
 
    bool file_new()
 		{ return new_flag; }
 
-   int get_blocklength()
+   size_t get_blocklength()
 		{ return blocklength; }
 
    size_t get_num_of_blocks()
@@ -93,13 +93,13 @@ public:
 
    int in_cache(int index);	// liefert Pos. im Cache oder -1
 
-   bool read_block(Block b,int i);
+   bool read_block(Block b,size_t i);
 
-   bool write_block(Block b,int i);
+   bool write_block(Block b,size_t i);
 
-   bool fix_block(int i);
+   bool fix_block(size_t i);
 
-   bool unfix_block(int i);
+   bool unfix_block(size_t i);
 
    void unfix_all();
